requirements.txt
Flask==2.3.2
python-dotenv==1.0.1
gunicorn==20.1.0
google-cloud-datastore
google-cloud-storage
python-jose[cryptography]
requests
